# api-spec/src/main/resources/openapi/product-sorter-v1.yaml
openapi: 3.1.0

info:
  title: Product Sorter API
  description: REST service that sorts a product catalog by weighted scoring criteria
  version: 1.0.0

paths:
  /products:
    get:
      operationId: getProducts
      tags:
        - products
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/sort:
    post:
      operationId: sortProducts
      tags:
        - products
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SortRequest'
      responses:
        '200':
          description: Products sorted by weighted criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '400':
          description: Invalid weights
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SortRequest:
      type: object
      required:
        - weights
      properties:
        weights:
          type: object
          description: Weight for each sorting criterion (e.g. salesUnits, stockRatio)
          additionalProperties:
            type: number
            format: double
            minimum: 0
            maximum: 1
      example:
        weights:
          salesUnits: 0.7
          stockRatio: 0.3

    ProductPageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        size:
          type: integer
        nextCursor:
          type: string

    ProductResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        salesUnits:
          type: integer
        stock:
          type: array
          items:
            $ref: '#/components/schemas/StockDto'
        score:
          type: number
          format: double
          nullable: true

    StockDto:
      type: object
      properties:
        size:
          type: string
        quantity:
          type: integer

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time